added post-dismissal option to UIActionSheet blocks

+++ UIKit/UIColor+SA_Additions.m
@@ -173,7 +173,7 @@ NSString *				NSStringFromCGColor(CGColorRef color) {
 	
 	if (componentCount == 2) return $S(@"White: %.0f, a: %.0f", comp[0], comp[1]);
 	
-	if (componentCount != 4) return $S(@"not an RGB color (%d comp)", (int) componentCount);
+	if (componentCount != 4) return $S(@"not an RGB color (%d comp)", (UInt16) componentCount);
 	return $S(@"R: %.0f, G: %.0f, B: %.0f, a: %.0f", comp[0], comp[1], comp[2], comp[3]);
 }
